requirements.txt
django==1.5.8
argparse==1.2.1
identity-toolkit-python-client>=0.1.6
oauth2client==1.4.5
ipython>=3.0.0